Backend: size_ha-Spaltentyp auf DOUBLE korrigieren (CI-Fix)
Hibernates Schema-Validierung (ddl-auto: validate) schlug in CI fehl:
FieldSnapshot.sizeHa ist ein Java "double", dafuer erwartet Hibernate den
SQL-Typ DOUBLE/float(53), nicht das in V4 verwendete DECIMAL(10,2).
V4 wurde noch nirgends angewendet (Teil dieses noch nicht gemergten PRs),
daher direkt korrigiert statt einer neuen Migration.

+++ backend/src/main/resources/db/migration/V4__create_field_snapshot_table.sql
@@ -9,8 +9,11 @@ CREATE TABLE field_snapshot (
     list_index          INT            NOT NULL,
     field_id            INT            NOT NULL,
     owner_farm_id       INT            NOT NULL,
-    size_ha             DECIMAL(10, 2) NOT NULL,
-    price               BIGINT         NOT NULL,
+    -- DOUBLE statt DECIMAL, da FieldSnapshot.sizeHa ein Java "double" ist -
+    -- Hibernates Schema-Validierung (ddl-auto: validate) erwartet dafuer
+    -- genau diesen SQL-Typ.
+    size_ha             DOUBLE NOT NULL,
+    price               BIGINT NOT NULL,
     PRIMARY KEY (id),
     CONSTRAINT fk_field_snapshot_world FOREIGN KEY (world_snapshot_id) REFERENCES world_snapshot (id) ON DELETE CASCADE,
     INDEX idx_field_snapshot_world (world_snapshot_id)
